프로젝트 수정하면 context api

+++ src/components/sidebar/project/list/item/manage/expel/Expel.jsx
@@ -1,25 +1,31 @@
+import { useContext } from "react";
+import { ProjectContext } from "@/context/ProjectContext";
 import styles from "./expel.module.css";
 
 const Expel = ({ userId, projectId }) => {
-  
+  const { refreshProjects } = useContext(ProjectContext);
+
   const expelHandler = async () => {
-    try {
-      const response = await fetch(`/api/user/expel`, {
-        method: "DELETE",
-        headers: {
-          "Content-Type": "application/json"
-        },
-        body: JSON.stringify({
-          userId: userId,
-          projectId: projectId
-        }),
-      });
-      if (!response.ok) {
-        throw new Error("추방 에러");
+    if (confirm("추방할?")) {
+      try {
+        const response = await fetch(`/api/user/expel`, {
+          method: "DELETE",
+          headers: {
+            "Content-Type": "application/json"
+          },
+          body: JSON.stringify({
+            userId: userId,
+            projectId: projectId
+          }),
+        });
+        if (!response.ok) {
+          throw new Error("추방 에러");
+        }
+        alert("내보내기 성공");
+        refreshProjects();
+      } catch (error) {
+        console.error(error);
       }
-      alert("내보내기 성공");
-    } catch (error) {
-      console.error(error);
     }
   }
